reorganze, try to avoid giant files
