+++ bracketcolors/src/Makefile.am
@@ -6,7 +6,11 @@ geanyplugins_LTLIBRARIES = bracketcolors.la
 bracketcolors_srcs = \
     bracketcolors.cc \
     BracketMap.cc \
-    BracketMap.h
+    BracketMap.h \
+    Utils.h \
+    Utils.cc \
+    Configuration.h \
+    Configuration.cc
 
 bracketcolors_la_SOURCES = $(bracketcolors_srcs)
 bracketcolors_la_CXXFLAGS = $(AM_CXXFLAGS) $(AM_CFLAGS) -DG_LOG_DOMAIN=\"BracketColors\"
